refactor: split input and central rendering; keep modules <500 lines
- Extract keyboard/mouse handling to src/input.rs
- Extract central panel rendering to src/central.rs
- Wire modules in main and app update; build verified
- app.rs now ~541 lines (down from 757); next step: move About and theme code if needed
