+++ src/app.rs
@@ -416,117 +416,8 @@ impl eframe::App for FileViewerApp {
 
         let mut file_to_load: Option<PathBuf> = None;
 
-        // Keyboard shortcuts
-        let mut toggle_dark = false;
-        ctx.input(|i| {
-            if i.modifiers.command && i.key_pressed(egui::Key::O) {
-                self.start_open_file_dialog();
-            }
-            if i.modifiers.command && i.key_pressed(egui::Key::D) {
-                toggle_dark = true;
-            }
-            if i.modifiers.command && i.key_pressed(egui::Key::F) {
-                self.search_active = true;
-            }
-            if i.modifiers.command && i.key_pressed(egui::Key::L) {
-                self.show_line_numbers = !self.show_line_numbers;
-                crate::settings::save_settings_to_disk(self);
-            }
-            if i.modifiers.command && i.key_pressed(egui::Key::W) {
-                self.word_wrap = !self.word_wrap;
-                crate::settings::save_settings_to_disk(self);
-            }
-
-            // Ctrl + Mouse wheel zoom for content
-            if i.modifiers.command && i.raw_scroll_delta.y != 0.0 {
-                let dir = i.raw_scroll_delta.y.signum();
-                match &self.content {
-                    Some(Content::Text(_)) => {
-                        let factor = if dir > 0.0 { 1.05 } else { 1.0 / 1.05 };
-                        self.text_zoom = (self.text_zoom * factor).clamp(0.6, 3.0);
-                    }
-                    Some(Content::Image(_)) => {
-                        self.image_fit = false;
-                        let factor = if dir > 0.0 { 1.10 } else { 1.0 / 1.10 };
-                        self.image_zoom = (self.image_zoom * factor).clamp(0.1, 6.0);
-                    }
-                    _ => {}
-                }
-            }
-
-            // Reset and keyboard zoom shortcuts
-            if i.modifiers.command && i.key_pressed(egui::Key::Num0) {
-                match &self.content {
-                    Some(Content::Text(_)) => self.text_zoom = 1.0,
-                    Some(Content::Image(_)) => { self.image_fit = false; self.image_zoom = 1.0; },
-                    _ => {}
-                }
-            }
-            if i.modifiers.command && i.key_pressed(egui::Key::Equals) {
-                match &self.content {
-                    Some(Content::Text(_)) => self.text_zoom = (self.text_zoom * 1.05).clamp(0.6, 3.0),
-                    Some(Content::Image(_)) => { self.image_fit = false; self.image_zoom = (self.image_zoom * 1.10).clamp(0.1, 6.0); },
-                    _ => {}
-                }
-            }
-            if i.modifiers.command && i.key_pressed(egui::Key::Minus) {
-                match &self.content {
-                    Some(Content::Text(_)) => self.text_zoom = (self.text_zoom / 1.05).clamp(0.6, 3.0),
-                    Some(Content::Image(_)) => { self.image_fit = false; self.image_zoom = (self.image_zoom / 1.10).clamp(0.1, 6.0); },
-                    _ => {}
-                }
-            }
-
-            // Navigation with arrow keys for current content type
-            if i.key_pressed(egui::Key::ArrowRight) {
-                if let Some(cur) = self.current_path.clone() {
-                    match self.content {
-                        Some(Content::Image(_)) => {
-                            if let Some(next) = crate::io::neighbor_image(&cur, true) { file_to_load = Some(next); }
-                        }
-                        Some(Content::Text(_)) => {
-                            if let Some(next) = crate::io::neighbor_text(&cur, true) { file_to_load = Some(next); }
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            if i.key_pressed(egui::Key::ArrowLeft) {
-                if let Some(cur) = self.current_path.clone() {
-                    match self.content {
-                        Some(Content::Image(_)) => {
-                            if let Some(prev) = crate::io::neighbor_image(&cur, false) { file_to_load = Some(prev); }
-                        }
-                        Some(Content::Text(_)) => {
-                            if let Some(prev) = crate::io::neighbor_text(&cur, false) { file_to_load = Some(prev); }
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            // Support '<' and '>' typed keys for both images and text
-            for ev in &i.events {
-                if let egui::Event::Text(t) = ev {
-                    if t == ">" {
-                        if let Some(cur) = self.current_path.clone() {
-                            match self.content {
-                                Some(Content::Image(_)) => { if let Some(next) = crate::io::neighbor_image(&cur, true) { file_to_load = Some(next); } }
-                                Some(Content::Text(_)) => { if let Some(next) = crate::io::neighbor_text(&cur, true) { file_to_load = Some(next); } }
-                                _ => {}
-                            }
-                        }
-                    } else if t == "<" {
-                        if let Some(cur) = self.current_path.clone() {
-                            match self.content {
-                                Some(Content::Image(_)) => { if let Some(prev) = crate::io::neighbor_image(&cur, false) { file_to_load = Some(prev); } }
-                                Some(Content::Text(_)) => { if let Some(prev) = crate::io::neighbor_text(&cur, false) { file_to_load = Some(prev); } }
-                                _ => {}
-                            }
-                        }
-                    }
-                }
-            }
-        });
+        // Keyboard + mouse input (delegated)
+        let toggle_dark = crate::input::handle_input(self, ctx, &mut file_to_load);
 
         // Poll background file open dialog
         if self.file_open_in_flight {
@@ -637,116 +528,9 @@ impl eframe::App for FileViewerApp {
             crate::ui::status_extra(ui, self);
         });
 
-        // Main Content
+        // Main Content (delegated)
         egui::CentralPanel::default().show(ctx, |ui| {
-            if let Some(err) = &self.error_message {
-                ui.colored_label(egui::Color32::RED, format!("Error: {}", err));
-            }
-
-            if let Some(content) = &self.content {
-                match content {
-                    Content::Text(text) => {
-                        let mut frame = egui::Frame::group(ui.style());
-                        frame.fill = self.code_theme.background();
-                        frame.show(ui, |ui| {
-                            // Wrap preference
-                            ui.style_mut().wrap_mode = Some(if self.word_wrap { egui::TextWrapMode::Wrap } else { egui::TextWrapMode::Extend });
-                            egui::ScrollArea::both().auto_shrink([false, false]).show(ui, |ui| {
-                                let text_style = egui::TextStyle::Monospace;
-                                let mut font_id = text_style.resolve(ui.style());
-                                font_id.size = (font_id.size * self.text_zoom).clamp(8.0, 48.0);
-                                let text_color = self.code_theme.foreground();
-
-                                let do_line_numbers = self.show_line_numbers && !self.text_is_big;
-                                let do_highlight = !self.text_is_big && text.len() <= HIGHLIGHT_CHAR_THRESHOLD;
-                                if do_line_numbers || do_highlight || !self.search_query.is_empty() {
-                                    let mut bracket_depth: i32 = 0;
-                                    let mut in_block_comment = false;
-                                    let ext = self
-                                        .current_path
-                                        .as_ref()
-                                        .and_then(|p| p.extension().and_then(|s| s.to_str()))
-                                        .unwrap_or("")
-                                        .to_lowercase();
-                                    // Determine target line for current match
-                                    let target_line_from_search = if !self.search_query.is_empty() && self.search_count > 0 {
-                                        search::find_target_line(text, &self.search_query, self.search_current)
-                                    } else { None };
-                                    // Render per line and capture rect
-                                    let mut counter: usize = 0;
-                                    let mut target_rect: Option<egui::Rect> = None;
-                                    for (i, line) in text.lines().enumerate() {
-                                        let mut line_job = LayoutJob::default();
-                                        if do_line_numbers {
-                                            line_job.append(&format!("{:>4} ", i + 1), 0.0, egui::TextFormat { font_id: font_id.clone(), color: self.code_theme.comment(), ..Default::default() });
-                                        }
-                                        highlight::append_highlighted(&mut line_job, line, &ext, &self.search_query, font_id.clone(), text_color, do_highlight, &mut bracket_depth, self.search_current, &mut counter, &mut in_block_comment, self.code_theme);
-                                        let resp = ui.label(line_job);
-                                        if target_line_from_search == Some(i) || self.scroll_target_line == Some(i) { target_rect = Some(resp.rect); }
-                                    }
-                                    if let Some(rect) = target_rect { ui.scroll_to_rect(rect, Some(egui::Align::Center)); }
-                                    self.scroll_target_line = None;
-                                } else {
-                                    ui.label(RichText::new(text).monospace().size(font_id.size));
-                                }
-                            });
-                        });
-                    }
-                    Content::Image(texture) => {
-                        let viewport = ui.available_size();
-                        egui::ScrollArea::both().show(ui, |ui| {
-                            ui.centered_and_justified(|ui| {
-                                let size = texture.size();
-                                let mut effective_zoom = self.image_zoom;
-                                if self.image_fit {
-                                    // Use the outer viewport size captured before the ScrollArea
-                                    let sx = if size[0] > 0 { viewport.x / size[0] as f32 } else { 1.0 };
-                                    let sy = if size[1] > 0 { viewport.y / size[1] as f32 } else { 1.0 };
-                                    let fit = sx.min(sy);
-                                    if fit.is_finite() && fit > 0.0 {
-                                        effective_zoom = fit.clamp(0.1, 6.0);
-                                    }
-                                }
-                                let desired = egui::vec2(size[0] as f32 * effective_zoom, size[1] as f32 * effective_zoom);
-                                let image = egui::Image::new(texture).fit_to_exact_size(desired);
-                                let resp = ui.add(image);
-                                if resp.hovered() {
-                                    let scroll = ui.input(|i| i.raw_scroll_delta.y);
-                                    if scroll != 0.0 {
-                                        self.image_fit = false;
-                                        let factor = if scroll > 0.0 { 1.10 } else { 1.0 / 1.10 };
-                                        self.image_zoom = (self.image_zoom * factor).clamp(0.1, 6.0);
-                                    }
-                                }
-                            });
-                        });
-                    }
-                }
-            } else if self.error_message.is_none() {
-                // Modern welcome screen
-                ui.vertical_centered(|ui| {
-                    ui.add_space(ui.available_height() * 0.2);
-                    
-                    // App icon and title
-                    ui.label(RichText::new("📁").size(64.0));
-                    ui.add_space(16.0);
-                    ui.label(RichText::new("Gemini File Viewer").heading().strong());
-                    ui.add_space(8.0);
-                    ui.label(RichText::new("A modern file viewer for text and images").weak());
-                    
-                    ui.add_space(32.0);
-                    
-                    // Centered Open File button
-                    ui.vertical_centered(|ui| {
-                        let mut open_button = egui::Button::new(RichText::new("📂 Open File").strong());
-                        open_button = open_button.fill(egui::Color32::from_rgb(34, 197, 94)); // Green
-                        if ui.add(open_button).clicked() {
-                            self.start_open_file_dialog();
-                        }
-                    });
-                    
-                });
-            }
+            crate::central::render_central_panel(ui, self);
         });
 
         // Deferred file loading to avoid borrow issues
